requirements.txt
openai>=1.0.0
typing-extensions>=4.0.0 
python-dotenv>=1.0.0